backend/requirements.txt
# ── Core Orchestration ────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-core==0.2.38
langgraph==0.1.19

# ── LLM Inference ─────────────────────────────────────────────────
groq==0.9.0

# ── Vector Store ──────────────────────────────────────────────────
chromadb==0.5.5

# ── Hybrid Retrieval + Reranking ──────────────────────────────────
rank-bm25==0.2.2
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0

# ── Document Ingestion ────────────────────────────────────────────
pypdf==4.3.1

# ── Web Search ────────────────────────────────────────────────────
tavily-python==0.3.3

# ── UI ────────────────────────────────────────────────────────────
gradio==4.42.0
fastapi==0.112.0
uvicorn==0.30.6

# ── Validation ────────────────────────────────────────────────────
pydantic==2.8.2

# ── Observability ─────────────────────────────────────────────────
langsmith==0.1.99

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv==1.0.1
tqdm==4.66.5

# ── Testing ───────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
